Add a Document struct for display
These come out of solr.
Record is just for ingesting.
