+++ lib/digital_collections/record.ex
@@ -1,6 +1,6 @@
 defmodule DigitalCollections.Record do
   @moduledoc """
-    A struct that holds our record schema
+    A struct that holds our record schema. Used for ingest
   """
   defstruct [
     :id,
@@ -18,18 +18,6 @@ defimpl DigitalCollections.Solr.Encoder, for: DigitalCollections.Record do
       titles_ssim: record.titles,
       authors_ssim: record.authors,
       subjects_ssim: record.subjects
-    }
-  end
-end
-
-defmodule DigitalCollections.Solr.Decoder do
-  # We may have different kind of records for ingest purposes, but anything coming out of solr will just be a Record, so we define the decoding here in this file.
-  def from_solr(map) do
-    %DigitalCollections.Record{
-      id: map["id"],
-      titles: map["titles_ssim"],
-      authors: map["authors_ssim"],
-      subjects: map["subjects_ssim"]
     }
   end
 end
