+++ test/digital_collections/catalog_test.exs
@@ -1,6 +1,6 @@
 defmodule CatalogTest do
   use ExUnit.Case
-  alias DigitalCollections.{Record, Catalog}
+  alias DigitalCollections.{Record, Catalog, Document}
 
   test ".add" do
     record = %Record{
@@ -11,7 +11,20 @@ defmodule CatalogTest do
     }
 
     {:ok, _response} = Catalog.add(record)
-    record = Catalog.get(record.id)
-    assert record.titles == ["My Book"]
+  end
+
+  test ".get" do
+    record = %Record{
+      id: "1",
+      titles: ["My Book"],
+      authors: ["Me"],
+      subjects: ["computers", "hacky sack"]
+    }
+
+    {:ok, _response} = Catalog.add(record)
+
+    document = Catalog.get(record.id)
+    %Document{} = document
+    assert document.titles == ["My Book"]
   end
 end
